requirements.txt
cx-Oracle==8.2.1
joblib==1.1.0
numpy==1.21.2
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
threadpoolctl==3.0.0
